% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawToStandard.R
\name{cohortToStandardCycle}
\alias{cohortToStandardCycle}
\title{RawToStandard
Raw data to standard data form for visualization}
\usage{
cohortToStandardCycle(connectionDetails, resultDatabaseSchema, cohortTable,
  targetCohortIds, cohortName, gapDate)
}
\arguments{
\item{tableType}{'cohort' or 'episode'}
}
\value{
standard data for visualization
}
\description{
RawToStandard
Raw data to standard data form for visualization
}
\keyword{standard}
