% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEpisodeCohort.R
\name{createCycleCohort}
\alias{createCycleCohort}
\title{CreateCohort
Create cohort of interest by using condition concept Ids.}
\usage{
createCycleCohort(createCohortTable = F, connectionDetails,
  oracleTempSchema = NULL, cdmDatabaseSchema, cohortDatabaseSchema,
  oncologyDatabaseSchema, vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortTable, episodeTable, conceptIdSet = c(),
  includeConceptIdSetDescendant = F, collapseGapSize = 0,
  targetCohortId)
}
\description{
CreateCohort
Create cohort of interest by using condition concept Ids.
}
