% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallRawData.R
\name{episodeCycle}
\alias{episodeCycle}
\title{CallRawData
Bring the chemotherapy raw data table.}
\usage{
episodeCycle(connectionDetails, vocaDatabaseSchema, oncologyDatabaseSchema,
  episodeTable, targetRegimenIds)
}
\value{
raw data for visualization
}
\description{
CallRawData
Bring the chemotherapy raw data table.
}
\keyword{data}
\keyword{raw}
