% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{parameterTree}
\alias{parameterTree}
\title{Visualization
Visualization tool for episode table.}
\usage{
parameterTree(targetRegimen)
}
\value{
Graph or table for episode table
}
\description{
Visualization
Visualization tool for episode table.
}
\keyword{graph,visualization}
