% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTree.R
\name{readJson}
\alias{readJson}
\title{Visualization
Visualization tool for episode table.}
\usage{
readJson(jsonName = "regimenDrugSettingJsonForm.json")
}
\value{
Graph or table for episode table
}
\description{
Visualization
Visualization tool for episode table.
}
\keyword{graph,visualization}
