% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YearlyGraph.R
\name{regimenInYear}
\alias{regimenInYear}
\title{Yearly graph
Yearly graph in treatment regimen}
\usage{
regimenInYear(connectionDetails, vocaDatabaseSchema,
  oncologyDatabaseSchema, episodeTable, targetRegimen, fromYear, toYear)
}
\value{
Yearly treatment regimen highcharter graph
}
\description{
Yearly graph
Yearly graph in treatment regimen
}
\keyword{year}
