% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YearlyGraph.R
\name{yearlyGraph}
\alias{yearlyGraph}
\title{Yearly graph
Yearly graph in treatment regimen}
\usage{
yearlyGraph(connectionDetails, vocaDatabaseSchema, targetCohortIds,
  fromYear, toYear)
}
\value{
Yearly treatment regimen highcharter graph
}
\description{
Yearly graph
Yearly graph in treatment regimen
}
\keyword{year}
