% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepetitionTrendHeatmap.R
\name{cohortToStandardCycle}
\alias{cohortToStandardCycle}
\title{Heatmap
Trends in regimen heatmap}
\usage{
cohortToStandardCycle(connectionDetails, resultDatabaseSchema, cohortTable,
  targetCohortIds, cohortName, gapDate)
}
\value{
repitition trend heatmap
}
\description{
Heatmap
Trends in regimen heatmap
}
\keyword{heatmap}
