% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEpisodeCohort.R
\name{createCycleCohort}
\alias{createCycleCohort}
\title{CreateCohort
Create cohort of interest by using condition concept Ids.}
\usage{
createCycleCohort(createCohortTable = F, connectionDetails,
  oracleTempSchema = NULL, cdmDatabaseSchema, cohortDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema, cohortTable,
  conceptIdSet = c(), includeConceptIdSetDescendant = F,
  targetCohortId)
}
\description{
CreateCohort
Create cohort of interest by using condition concept Ids.
}
